Workaround incompatibility between dependencies
Before this change, `cargo install --path .` fails with the following
error:
```
error[E0433]: failed to resolve: could not find `__rt` in `quote`
   --> /Users/.../.cargo/registry/src/github.com-1ecc6299db9ec823/failure_derive-0.1.6/src/lib.rs:107:70
    |
107 | fn display_body(s: &synstructure::Structure) -> Result<Option<quote::__rt::TokenStream>, Error> {
    |                                                                      ^^^^ could not find `__rt` in `quote`

error: aborting due to previous error

For more information about this error, try `rustc --explain E0433`.
error: Could not compile `failure_derive`.
```

An explanation for this error and a hint for the workaround are
available here:
https://users.rust-lang.org/t/failure-derive-compilation-error/39062